chore: add video

+++ domain/chat_invite_link.go
@@ -13,7 +13,7 @@ type ChatInviteLink struct {
 	// IsRevoked true, if the link is revoked.
 	IsRevoked bool `json:"is_revoked"`
 	// ExpireDate point in time (Unix timestamp) when the link will expire or has been expired.
-	ExpireDate int64 `json:"expire_date"`
+	ExpireDate int `json:"expire_date"`
 	// MemberLimit maximum number of users that can be members of the chat
 	// simultaneously after joining the chat via this invite link; 1-99999.
 	MemberLimit int `json:"member_limit"`
